feat: refactored DatabaseLogger and associated classes to make it cleaner and compatible with subclassing, added pef entry, and triaging

+++ Starter Application/app/src/main/java/com/example/smartAir/triaging/RecentHistoryEntry.java
@@ -44,7 +44,7 @@ public class RecentHistoryEntry extends Fragment {
             public void onTextChanged(CharSequence s, int start, int before, int count) {
 
                 if (!s.isEmpty()) {
-                    triageController.setPEF(Float.parseFloat(s.toString()));
+                    triageController.setPEF(Double.parseDouble(s.toString()));
                 } else {
                     triageController.setPEF(null);
                 }
