+++ Starter Application/app/src/main/java/com/example/smartAir/triaging/TriageLogEntryData.java
@@ -3,12 +3,12 @@ package com.example.smartAir.triaging;
 import com.example.smartAir.data.DatabaseLogEntryData;
 
 public class TriageLogEntryData extends DatabaseLogEntryData {
-    public Float pef;
+    public Double pef;
     public Integer rescueCount;
     public String triageDecision;
     public String guidanceProvided;
 
-    public TriageLogEntryData(Float pef, Integer rescueCount, String triageDecision, String guidanceProvided) {
+    public TriageLogEntryData(Double pef, Integer rescueCount, String triageDecision, String guidanceProvided) {
         this.pef = pef;
         this.rescueCount = rescueCount;
         this.triageDecision = triageDecision;
@@ -17,7 +17,8 @@ public class TriageLogEntryData extends DatabaseLogEntryData {
 
     public TriageLogEntryData() {} // Used by DatabaseEntryData
 
-    public Float getPef() {
+
+    public Double getPef() {
         return pef;
     }
